move memecrypto into submodule
